add initial random seed configuration

+++ src/Backdash/Network/Messages/SyncRequest.cs
@@ -7,7 +7,7 @@ namespace Backdash.Network.Messages;
 [Serializable, StructLayout(LayoutKind.Sequential, Pack = 4)]
 record struct SyncRequest : IUtf8SpanFormattable
 {
-    public uint RandomRequest; /* please reply with this random data */
+    public ushort RandomRequest;
     public long Ping;
 
     public readonly void Serialize(in BinaryRawBufferWriter writer)
@@ -18,7 +18,7 @@ record struct SyncRequest : IUtf8SpanFormattable
 
     public void Deserialize(in BinaryBufferReader reader)
     {
-        RandomRequest = reader.ReadUInt32();
+        RandomRequest = reader.ReadUInt16();
         Ping = reader.ReadInt64();
     }
